created schema for /users, still need to abstract schemas

+++ apps/backend/src/routes/user.ts
@@ -5,18 +5,22 @@ import {
     UserRequestParams,
     UserWalletsRequestBody
 } from "../types/routeParams";
-import userSchema from "../../docs/user.json";
-import { logError } from "../utils/utils";
+import UserSchema from "../../docs/schemas/user.json";
+import { logError } from "../lib/utils";
 
 const user = (server: FastifyInstance, _: any, done: () => void) => {
     const { prisma } = server;
 
-    const { userIdSchema, walletsSchema } = userSchema;
+    const { GetUserSchema, PostUserWalletSchema, DeleteUserWalletSchema } =
+        UserSchema;
 
-    server.get("/:userId", userIdSchema, async (request, reply) => {
+    server.get("/:userId", GetUserSchema, async (request, reply) => {
         const { userId } = request.params as UserRequestParams;
 
-        if (!userId) logError(reply, 400, "missing user id param");
+        if (!userId) {
+            logError(reply, 400, "missing user id param");
+            return;
+        }
 
         try {
             const user = await prisma.user.findUnique({
@@ -34,16 +38,23 @@ const user = (server: FastifyInstance, _: any, done: () => void) => {
         }
     });
 
-    server.post("/:userId/wallets", walletsSchema, async (request, reply) => {
-        const { userId } = request.params as UserRequestParams;
+    server.post(
+        "/:userId/wallets",
+        PostUserWalletSchema,
+        async (request, reply) => {
+            const { userId } = request.params as UserRequestParams;
+            if (!userId) {
+                logError(reply, 400, "missing user id param");
+                return;
+            }
 
-        if (!userId) logError(reply, 400, "missing user id param");
+            const { walletAddress } = request.body as UserWalletsRequestBody;
+            if (!walletAddress) {
+                logError(reply, 400, "missing wallet address param");
+                return;
+            }
 
-        const { walletAddress, operation } =
-            request.body as UserWalletsRequestBody;
-
-        try {
-            if (operation === "link") {
+            try {
                 const wallet = await prisma.wallet.create({
                     data: {
                         address: walletAddress,
@@ -52,20 +63,45 @@ const user = (server: FastifyInstance, _: any, done: () => void) => {
                 });
 
                 return wallet;
+            } catch (e) {
+                if (e instanceof PrismaClientKnownRequestError)
+                    logError(reply, 500, e.message);
+
+                logError(reply, 500, "creating wallet");
+            }
+        }
+    );
+
+    server.delete(
+        "/:userId/wallets",
+        DeleteUserWalletSchema,
+        async (request, reply) => {
+            const { userId } = request.params as UserRequestParams;
+            if (!userId) {
+                logError(reply, 400, "missing user id param");
+                return;
             }
 
-            await prisma.wallet.delete({
-                where: {
-                    address: walletAddress
-                }
-            });
-        } catch (e) {
-            if (e instanceof PrismaClientKnownRequestError)
-                logError(reply, 500, e.message);
+            const { walletAddress } = request.body as UserWalletsRequestBody;
+            if (!walletAddress) {
+                logError(reply, 400, "missing wallet address param");
+                return;
+            }
 
-            logError(reply, 500, "creating wallet");
+            try {
+                await prisma.wallet.delete({
+                    where: {
+                        address: walletAddress
+                    }
+                });
+            } catch (e) {
+                if (e instanceof PrismaClientKnownRequestError)
+                    logError(reply, 500, e.message);
+
+                logError(reply, 500, "deleting wallet");
+            }
         }
-    });
+    );
 
     done();
 };
